Refactor edit.blade.php to include reset password checkbox
This commit modifies the edit.blade.php file to include a reset password checkbox in the form. It adds a new form group with a checkbox input and label for the "Reset Password" functionality. This change enhances the edit view by allowing users to easily reset the password of a student.