Refactor ClassRoomController store and update methods to redirect to the show page
This commit refactors the store and update methods in the ClassRoomController to redirect to the show page of the created or updated classroom. Previously, both methods redirected to the index page. By redirecting to the show page, users can immediately view the details of the created or updated classroom. This change improves the user experience and enhances the functionality of the application.