Refactor code to add 'ip_address', 'user_agent', and 'location' fields to StudentPresent model
Add the 'ip_address', 'user_agent', and 'location' fields to the StudentPresent model. These fields are now fillable and will be used to store the IP address, user agent, and location information of the student when they mark their attendance. This change enhances the data captured for each attendance record and improves the overall functionality of the application.